스타일 수정

+++ components/PostCardBig/CommentBar.js
@@ -17,6 +17,8 @@ const editComment = () => async() => {
 
 const deleteComment = (target) => async () => {
   try {
+    const isDelete = confirm('정말로 삭제하시겠습니까?');
+    if (!isDelete) return;
     const result = await axios.delete(`/comment?commentId=${target.id}`);
     alert(result.data);
   } catch (error) {
